third_party/WebKit/LayoutTests/http/tests/fetch/window/pageimportancesignals.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<script>
description("PageImportanceSignals should be recorded after script initiated non-GET fetch.");
window.jsTestIsAsync = true;

function parseJSON(response) {
  return response.json;
}

function test1() {
  console.log("Testing script initiated GET fetch.");
  window.fetch("../resources/simple.json")
    .then(parseJSON)
    .then(function(data) {
      if (window.testRunner)
          window.testRunner.dumpPageImportanceSignals();

      test2();
    })
    .catch(function(error) {
      testFailed("fetch failed: "+error);
      finishJSTest();
    });
}

function test2() {
  console.log("Testing script initiated POST fetch.");
  window.fetch("../resources/simple.json", { method: "POST" })
    .then(parseJSON)
    .then(function(data) {
      if (window.testRunner)
          window.testRunner.dumpPageImportanceSignals();

      finishJSTest();
    })
    .catch(function() {
      testFailed("fetch failed: "+error);
      finishJSTest();
    });
}

test1();
</script>
