third_party/WebKit/LayoutTests/http/tests/fetch/window/pageimportancesignals-expected.txt

CONSOLE MESSAGE: line 12: Testing script initiated GET fetch.
WebPageImportanceSignals:
  hadFormInteraction: false
  issuedNonGetFetchFromScript: false
CONSOLE MESSAGE: line 28: Testing script initiated POST fetch.
WebPageImportanceSignals:
  hadFormInteraction: false
  issuedNonGetFetchFromScript: true
PageImportanceSignals should be recorded after script initiated non-GET fetch.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
